Expansion algorithm complete
The expansion algorithm is now fully implemented. It differs from the current version of the spec in that it expands @list-objects to arrays and that keyword aliases are not supported within a context.

The expansion of @graph might need to be changed to keep the @graph keyword (see https://github.com/json-ld/json-ld.org/issues/94).