Don't drop @graph if it's not the only property at the top-level's object
It was decided that "If @graph is the only property in the document's top level object, it MUST be dropped in expansion and compaction. In all other cases (which includes @graph at the document's top level object when there are other properties other than @context at the same level), @graph MUST NOT be dropped in expansion and compaction." Updated the processor accordingly.

See https://github.com/json-ld/json-ld.org/issues/99.