Implement compaction for property generators
This doesn't close the issue yet as currently only the best matching property generator is evaluated. Other, less specific property generators might apply if the most specific one doesn't. The problem is there's no way to know that before comparing the values for all of the property generator's IRIs.

This addresses #17.